Minor fixes for MSVC, fix server without clip control
(NOTE: Submodules ffl, ninTexUtils, and rio are not updated with the changes in 2024-07-08-msvc-fixes yet)
* src/Shader.cpp: Flip texCoords in-place rather than with a new buffer.
* include/mii_ext_MiiPort.h: Remove ver3StoreData because MSVC isn't compatible with its unions, and we don't need it anyway, we have FFLiMiiDataCore.
* src/RootTask.cpp: Replace ssize_t with int. When there is no clip control, flip Y in the projection matrix and set front face culling from CCW to CW. Give back control to the window when finished with the render buffer.

+++ src/RootTask.cpp
@@ -387,7 +387,7 @@ void RootTask::calc_()
     if (mSocketIsListening &&
         (new_socket = accept(server_fd, (struct sockaddr *)&address, (socklen_t*)&addrlen)) > 0) {
         // Assuming data directly received is FFLStoreData
-        ssize_t read_bytes =
+        int read_bytes =
             //read(new_socket, &storeData, sizeof(FFLStoreData));
             // NOTE: this will store BOTH charinfo AND storedata so uh
             recv(new_socket, buf,
